Add functionality to add and remove task
