+++ src/index.html
@@ -20,7 +20,7 @@
            <button id="addButton"> <i class="fa fa-caret-right"></i></button>
         </div>
        
-        <hr>
+        <hr id="hrId">
         <ul id="list">
             
         </ul>
