+++ src/index.js
@@ -1,90 +1,34 @@
 import _ from 'lodash';
 import './style.css';
 
-class Task {
-    constructor(array) {
-        this.taskArray = array;
-    }
-
-    taskObject(addTask) {
-        const eachTask = {
-            description : addTask,
-            completed : false,
-            index: `${this.taskArray.length + 1}`
-        }
-        this.taskArray.push(eachTask);
-    }
-
-    removeTask(ptask) {
-        this.taskArray = this.taskArray.filter((task) => task !== ptask);
-    }
-}
-
-const addedTask = new Task([])
-
-
+import addedTask from "./module/class.js"
+import { addToLocalStorage, getFromLocalStorage } from "./module/localstorage.js"
+import displayList from "./module/displaylist.js"
 
 const filledTask = document.getElementById('addButton')
 filledTask.addEventListener('click', () => {
-const myTask = document.getElementById('addTask')
-addedTask.taskObject(myTask.value)
-addToLocalStorage();
-displayList()
-myTask.value = '';
+    const myTask = document.getElementById('addTask')
+    addedTask.taskObject(myTask.value)
+    addedTask.reAssignIndex()
+    displayList()
+    myTask.value = '';
 })
 
-
-const displayList = () => {
-const list = document.getElementById('list')
-list.innerHTML = '';
-if (addedTask.taskArray.length) {
-    for (let i = 0; i < addedTask.taskArray.length; i+=1){
-        const li = document.createElement('div')
-        li.classList.add('eachTask')
-        list.appendChild(li)
-
-        const taskInput = document.createElement('div')
-        taskInput.innerHTML = `<input type="checkbox" name=" " class="task">
-        <label for="">${addedTask.taskArray[i].description}</label>`
-        li.appendChild(taskInput)
-
-
-        const deleteButton = document.createElement('button')
-        deleteButton.classList.add('delButton')
-        deleteButton.innerHTML = `<i class="fa fa-ellipsis-v">`
-        
-        deleteButton.onclick = () => {
-            addedTask.removeTask(addedTask.taskArray[i]);
-            
-            addToLocalStorage();
-            displayList();
-        }
-
-        li.appendChild(deleteButton)    
+const handleInputChange = (id) => {
+   const theid = document.getElementById(`edible${id}`);
+   for(let p of addedTask.taskArray){
+    if(p.index === id){
+        p.description = theid.value
     }
-} else {
-    const noTask = document.createElement('p');
-    noTask.classList.add('no_task');
-    noTask.textContent = `No tasks added yet`;
-    list.appendChild(noTask)
-  }
-}
-
-
-
-const addToLocalStorage = () => {
-const stringifyArray = JSON.stringify(addedTask.taskArray);
-localStorage.setItem('storedTask', stringifyArray);
-}
-
-const getFromLocalStorage = () => {
-const stringifyArray = localStorage.getItem('storedTask');
-addedTask.taskArray = JSON.parse(stringifyArray);
-displayList()
+   }
+   addToLocalStorage()
+   displayList()
 }
 
 if (localStorage.getItem('storedTask') == null) {
-addToLocalStorage();
+    addToLocalStorage();
 } else {
-getFromLocalStorage();
+    getFromLocalStorage();
 }
+
+window.handleInputChange = handleInputChange;
